+++ tcex/tcex_batch_v2.py
@@ -919,7 +919,7 @@ class TcExBatch(object):
         return self._indicators_shelf
 
     def intrusion_set(self, name, xid=True):
-        """Add IntrusionSet data to Batch object.
+        """Add Intrusion Set data to Batch object.
 
         Args:
             name (str): The name for this Group.
@@ -1623,6 +1623,16 @@ class TcExBatch(object):
 
 class Group(object):
     """ThreatConnect Batch Group Object"""
+    __slots__ = [
+        '_attributes',
+        '_file_content',
+        '_group_data',
+        '_labels',
+        '_name',
+        '_processed',
+        '_type',
+        '_tags',
+        '_utils']
 
     def __init__(self, group_type, name, xid=True):
         """Initialize Class Properties.
@@ -1762,6 +1772,17 @@ class Group(object):
         return self._group_data
 
     @property
+    def date_added(self):
+        """Return Group dateAdded."""
+        return self._group_data.get('dateAdded')
+
+    @date_added.setter
+    def date_added(self, date_added):
+        """Set Indicator dateAdded."""
+        self._group_data['dateAdded'] = self._utils.format_datetime(
+            date_added, date_format='%Y-%m-%dT%H:%M:%SZ')
+
+    @property
     def file_data(self):
         """Return Group file (only supported for Document and Report)."""
         return {
@@ -1842,12 +1863,13 @@ class Group(object):
         return self._group_data.get('xid')
 
     def __str__(self):
-        """Return string represtentation of object."""
+        """Return string representation of object."""
         return json.dumps(self.data, indent=4)
 
 
 class Adversary(Group):
     """ThreatConnect Batch Adversary Object"""
+    __slots__ = []
 
     def __init__(self, name, xid=True):
         """Initialize Class Properties.
@@ -1861,6 +1883,7 @@ class Adversary(Group):
 
 class Campaign(Group):
     """ThreatConnect Batch Campaign Object"""
+    __slots__ = []
 
     def __init__(self, name, first_seen=None, xid=True):
         """Initialize Class Properties.
@@ -1889,6 +1912,7 @@ class Campaign(Group):
 
 class Document(Group):
     """ThreatConnect Batch Document Object"""
+    __slots__ = ['_file_data', '_group_data']
 
     def __init__(self, name, file_name, file_content=None, malware=False, password=None,
                  xid=True):
@@ -1955,6 +1979,7 @@ class Document(Group):
 
 class Email(Group):
     """ThreatConnect Batch Email Object"""
+    __slots__ = []
 
     def __init__(self, name, subject, header, body, to_addr=None, from_addr=None, xid=True):
         """Initialize Class Properties.
@@ -2011,6 +2036,7 @@ class Email(Group):
 
 class Event(Group):
     """ThreatConnect Batch Event Object"""
+    __slots__ = []
 
     def __init__(self, name, event_date=None, status=None, xid=True):
         """Initialize Class Properties.
@@ -2058,6 +2084,7 @@ class Event(Group):
 
 class Incident(Group):
     """ThreatConnect Batch Incident Object"""
+    __slots__ = []
 
     def __init__(self, name, event_date=None, status=None, xid=True):
         """Initialize Class Properties.
@@ -2122,6 +2149,7 @@ class Incident(Group):
 
 class IntrusionSet(Group):
     """ThreatConnect Batch Adversary Object"""
+    __slots__ = []
 
     def __init__(self, name, xid=True):
         """Initialize Class Properties.
@@ -2135,6 +2163,7 @@ class IntrusionSet(Group):
 
 class Report(Group):
     """ThreatConnect Batch Report Object"""
+    __slots__ = []
 
     def __init__(self, name, file_name=None, file_content=None, publish_date=None, xid=True):
         """Initialize Class Properties.
@@ -2188,6 +2217,7 @@ class Report(Group):
 
 class Signature(Group):
     """ThreatConnect Batch Signature Object"""
+    __slots__ = []
 
     def __init__(self, name, file_name, file_type, file_text, xid=True):
         """Initialize Class Properties.
@@ -2218,6 +2248,7 @@ class Signature(Group):
 
 class Threat(Group):
     """ThreatConnect Batch Threat Object"""
+    __slots__ = []
 
     def __init__(self, name, xid=True):
         """Initialize Class Properties.
@@ -2236,6 +2267,16 @@ class Threat(Group):
 
 class Indicator(object):
     """ThreatConnect Batch Indicator Object"""
+    __slots__ = [
+        '_attributes',
+        '_file_actions',
+        '_indicator_data',
+        '_labels',
+        '_occurrences',
+        '_summary',
+        '_tags',
+        '_type',
+        '_utils']
 
     def __init__(self, indicator_type, summary, rating=None, confidence=None, xid=True):
         """Initialize Class Properties.
@@ -2420,6 +2461,28 @@ class Indicator(object):
                     self._indicator_data['tag'].append(tag.data)
         return self._indicator_data
 
+    @property
+    def date_added(self):
+        """Return Indicator dateAdded."""
+        return self._indicator_data.get('dateAdded')
+
+    @date_added.setter
+    def date_added(self, date_added):
+        """Set Indicator dateAdded."""
+        self._indicator_data['dateAdded'] = self._utils.format_datetime(
+            date_added, date_format='%Y-%m-%dT%H:%M:%SZ')
+
+    @property
+    def last_modified(self):
+        """Return Indicator lastModified."""
+        return self._indicator_data.get('lastModified')
+
+    @last_modified.setter
+    def last_modified(self, last_modified):
+        """Set Indicator lastModified."""
+        self._indicator_data['lastModified'] = self._utils.format_datetime(
+            last_modified, date_format='%Y-%m-%dT%H:%M:%SZ')
+
     def occurrence(self, file_name=None, path=None, date=None):
         """Add a file Occurrence.
 
@@ -2525,6 +2588,7 @@ class Indicator(object):
 
 class Address(Indicator):
     """ThreatConnect Batch Address Object"""
+    __slots__ = []
 
     def __init__(self, ip, rating=None, confidence=None, xid=True):
         """Initialize Class Properties.
@@ -2540,6 +2604,7 @@ class Address(Indicator):
 
 class ASN(Indicator):
     """ThreatConnect Batch ASN Object."""
+    __slots__ = []
 
     def __init__(self, as_number, rating=None, confidence=None, xid=True):
         """Initialize Class Properties.
@@ -2555,6 +2620,7 @@ class ASN(Indicator):
 
 class CIDR(Indicator):
     """ThreatConnect Batch CIDR Object"""
+    __slots__ = []
 
     def __init__(self, block, rating=None, confidence=None, xid=True):
         """Initialize Class Properties.
@@ -2570,6 +2636,7 @@ class CIDR(Indicator):
 
 class EmailAddress(Indicator):
     """ThreatConnect Batch EmailAddress Object"""
+    __slots__ = []
 
     def __init__(self, address, rating=None, confidence=None, xid=True):
         """Initialize Class Properties.
@@ -2585,6 +2652,7 @@ class EmailAddress(Indicator):
 
 class File(Indicator):
     """ThreatConnect Batch File Object"""
+    __slots__ = []
 
     def __init__(self, md5=None, sha1=None, sha256=None, size=None, rating=None,
                  confidence=None, xid=True):
@@ -2655,6 +2723,7 @@ class File(Indicator):
 
 class Host(Indicator):
     """ThreatConnect Batch Host Object"""
+    __slots__ = []
 
     def __init__(self, hostname, dns_active=False, whois_active=False, rating=None,
                  confidence=None, xid=True):
@@ -2698,6 +2767,7 @@ class Host(Indicator):
 
 class Mutex(Indicator):
     """ThreatConnect Batch Mutex Object"""
+    __slots__ = []
 
     def __init__(self, mutex, rating=None, confidence=None, xid=True):
         """Initialize Class Properties.
@@ -2713,6 +2783,7 @@ class Mutex(Indicator):
 
 class RegistryKey(Indicator):
     """ThreatConnect Batch Registry Key Object"""
+    __slots__ = []
 
     def __init__(self, key_name, value_name, value_type, rating=None, confidence=None,
                  xid=True):
@@ -2732,6 +2803,7 @@ class RegistryKey(Indicator):
 
 class URL(Indicator):
     """ThreatConnect Batch URL Object"""
+    __slots__ = []
 
     def __init__(self, text, rating=None, confidence=None, xid=True):
         """Initialize Class Properties.
@@ -2747,6 +2819,7 @@ class URL(Indicator):
 
 class UserAgent(Indicator):
     """ThreatConnect Batch User Agent Object"""
+    __slots__ = []
 
     def __init__(self, text, rating=None, confidence=None, xid=True):
         """Initialize Class Properties.
@@ -2767,6 +2840,7 @@ class UserAgent(Indicator):
 
 class Attribute(object):
     """ThreatConnect Batch Attribute Object"""
+    __slots__ = ['_attribute_data', '_valid']
 
     def __init__(self, attr_type, attr_value, displayed=False, source=None, formatter=None):
         """Initialize Class Properties.
@@ -2844,6 +2918,7 @@ class Attribute(object):
 
 class FileAction(object):
     """ThreatConnect Batch FileAction Object"""
+    __slots__ = ['_action_data', '_children', 'xid']
 
     def __init__(self, parent_xid, relationship):
         """Initialize Class Properties.
@@ -2882,6 +2957,7 @@ class FileAction(object):
 
 class FileOccurrence(object):
     """ThreatConnect Batch FileAction Object."""
+    __slots__ = ['_occurrence_data', '_utils']
 
     def __init__(self, file_name=None, path=None, date=None):
         """Initialize Class Properties
@@ -2944,6 +3020,7 @@ class FileOccurrence(object):
 
 class SecurityLabel(object):
     """ThreatConnect Batch SecurityLabel Object."""
+    __slots__ = ['_label_data']
 
     def __init__(self, name, description=None, color=None):
         """Initialize Class Properties.
@@ -2997,6 +3074,7 @@ class SecurityLabel(object):
 
 class Tag(object):
     """ThreatConnect Batch Tag Object"""
+    __slots__ = ['_tag_data', '_valid']
 
     def __init__(self, name, formatter=None):
         """Initialize Class Properties.